fix: 수익률 소수점이하 미계산 결함 해결

+++ src/main/java/lotto/service/WinningsCalculationService.java
@@ -31,7 +31,7 @@ public class WinningsCalculationService {
             winningPrice += calculateWinningNumber(lotto, winningsNumber, bonusNumber);
         }
 
-        int revenue = winningPrice / user.getPaymentAmount() * 100;
+        double revenue = ((double) winningPrice / user.getPaymentAmount()) * 100;
         SystemGuidePrinter.winningHistoryMessage(winningResult, revenue);
     }
